AI Weather Simulator Agent/requirements.txt
Flask==2.3.3
requests==2.31.0
matplotlib==3.7.2
plotly==5.15.0
pandas==2.0.3
openai==0.28.0
python-dotenv==1.0.0
meteostat==1.6.7
numpy==1.24.3
